Create dummy variables for all categorical/factor variables in a data.frame

Create dummy/indicator variables for all categorical variables in a data.frame. Can be used as a pre-processing step before calling other functions within the package.

Usage

dummy_vars(df)

Arguments

df

a data.frame on which to make dummy variables for each categorical/factor variable, based on contrasts.

Value

a data.frame matching df but where where each categorical variable in df is replaced with indicator variables. All combinations of the indicator/dummy variable are returned. Naming convention of the new dummy variables is variable_level. For example, a factor variable in df named "colour" with levels "red", "green" and "purple" will be replaced with three columns (the new dummy variables), named colour_red, colour_green and colour_purple.

See also

Examples

dummy_vars(data.frame("Colour" = factor(sample(c("red",
                                                 "azure",
                                                 "green",
                                                 "white"),
                                              500,
                                              replace = TRUE))))
#>     Colour_azure Colour_green Colour_red Colour_white
#> 1              0            0          0            1
#> 2              0            0          1            0
#> 3              1            0          0            0
#> 4              0            0          0            1
#> 5              0            0          1            0
#> 6              0            0          0            1
#> 7              1            0          0            0
#> 8              1            0          0            0
#> 9              0            1          0            0
#> 10             0            0          1            0
#> 11             1            0          0            0
#> 12             0            0          0            1
#> 13             0            0          1            0
#> 14             1            0          0            0
#> 15             1            0          0            0
#> 16             0            0          1            0
#> 17             0            0          0            1
#> 18             0            0          1            0
#> 19             0            0          1            0
#> 20             1            0          0            0
#> 21             1            0          0            0
#> 22             0            1          0            0
#> 23             0            0          0            1
#> 24             0            0          1            0
#> 25             0            0          0            1
#> 26             0            0          1            0
#> 27             0            0          0            1
#> 28             1            0          0            0
#> 29             1            0          0            0
#> 30             1            0          0            0
#> 31             1            0          0            0
#> 32             0            0          1            0
#> 33             0            0          1            0
#> 34             0            0          0            1
#> 35             0            1          0            0
#> 36             1            0          0            0
#> 37             1            0          0            0
#> 38             1            0          0            0
#> 39             0            1          0            0
#> 40             0            0          1            0
#> 41             0            0          1            0
#> 42             1            0          0            0
#> 43             1            0          0            0
#> 44             0            0          1            0
#> 45             0            1          0            0
#> 46             0            1          0            0
#> 47             0            0          1            0
#> 48             0            0          0            1
#> 49             0            0          0            1
#> 50             0            0          0            1
#> 51             0            0          0            1
#> 52             0            1          0            0
#> 53             1            0          0            0
#> 54             0            0          0            1
#> 55             0            0          1            0
#> 56             0            0          1            0
#> 57             1            0          0            0
#> 58             0            1          0            0
#> 59             0            0          1            0
#> 60             1            0          0            0
#> 61             0            0          0            1
#> 62             1            0          0            0
#> 63             0            1          0            0
#> 64             0            0          0            1
#> 65             0            0          0            1
#> 66             0            1          0            0
#> 67             0            0          1            0
#> 68             0            0          1            0
#> 69             1            0          0            0
#> 70             1            0          0            0
#> 71             0            0          0            1
#> 72             0            0          1            0
#> 73             0            0          0            1
#> 74             0            0          1            0
#> 75             0            0          1            0
#> 76             1            0          0            0
#> 77             0            0          0            1
#> 78             1            0          0            0
#> 79             0            0          1            0
#> 80             0            0          1            0
#> 81             1            0          0            0
#> 82             0            1          0            0
#> 83             0            1          0            0
#> 84             0            0          0            1
#> 85             0            0          0            1
#> 86             0            1          0            0
#> 87             0            1          0            0
#> 88             0            0          0            1
#> 89             1            0          0            0
#> 90             1            0          0            0
#> 91             0            1          0            0
#> 92             1            0          0            0
#> 93             0            1          0            0
#> 94             0            0          1            0
#> 95             1            0          0            0
#> 96             0            0          1            0
#> 97             0            1          0            0
#> 98             0            0          1            0
#> 99             0            0          0            1
#> 100            0            1          0            0
#> 101            0            1          0            0
#> 102            0            0          0            1
#> 103            0            0          1            0
#> 104            0            1          0            0
#> 105            1            0          0            0
#> 106            1            0          0            0
#> 107            0            0          0            1
#> 108            0            0          1            0
#> 109            0            1          0            0
#> 110            0            1          0            0
#> 111            0            0          1            0
#> 112            0            1          0            0
#> 113            0            0          0            1
#> 114            1            0          0            0
#> 115            0            0          1            0
#> 116            0            1          0            0
#> 117            0            0          1            0
#> 118            1            0          0            0
#> 119            0            0          0            1
#> 120            0            1          0            0
#> 121            0            1          0            0
#> 122            0            0          1            0
#> 123            0            0          1            0
#> 124            0            0          1            0
#> 125            0            0          0            1
#> 126            0            1          0            0
#> 127            0            0          1            0
#> 128            0            0          1            0
#> 129            0            1          0            0
#> 130            1            0          0            0
#> 131            1            0          0            0
#> 132            0            1          0            0
#> 133            0            0          0            1
#> 134            0            1          0            0
#> 135            0            1          0            0
#> 136            1            0          0            0
#> 137            0            0          0            1
#> 138            1            0          0            0
#> 139            0            0          0            1
#> 140            1            0          0            0
#> 141            1            0          0            0
#> 142            0            1          0            0
#> 143            0            0          0            1
#> 144            1            0          0            0
#> 145            0            0          0            1
#> 146            0            0          1            0
#> 147            1            0          0            0
#> 148            0            0          1            0
#> 149            1            0          0            0
#> 150            0            0          0            1
#> 151            0            0          1            0
#> 152            0            1          0            0
#> 153            0            0          0            1
#> 154            1            0          0            0
#> 155            1            0          0            0
#> 156            0            0          1            0
#> 157            0            1          0            0
#> 158            0            1          0            0
#> 159            0            0          0            1
#> 160            0            0          0            1
#> 161            1            0          0            0
#> 162            0            0          1            0
#> 163            0            0          1            0
#> 164            0            0          0            1
#> 165            0            0          1            0
#> 166            0            0          1            0
#> 167            1            0          0            0
#> 168            0            0          1            0
#> 169            0            0          1            0
#> 170            0            1          0            0
#> 171            1            0          0            0
#> 172            0            0          0            1
#> 173            0            1          0            0
#> 174            0            1          0            0
#> 175            0            0          0            1
#> 176            1            0          0            0
#> 177            0            0          1            0
#> 178            0            0          0            1
#> 179            0            0          0            1
#> 180            0            1          0            0
#> 181            0            0          0            1
#> 182            1            0          0            0
#> 183            0            0          0            1
#> 184            0            0          1            0
#> 185            1            0          0            0
#> 186            0            1          0            0
#> 187            0            0          0            1
#> 188            0            0          1            0
#> 189            1            0          0            0
#> 190            0            0          0            1
#> 191            0            1          0            0
#> 192            1            0          0            0
#> 193            0            1          0            0
#> 194            0            0          1            0
#> 195            0            1          0            0
#> 196            0            1          0            0
#> 197            1            0          0            0
#> 198            0            0          0            1
#> 199            0            1          0            0
#> 200            0            0          1            0
#> 201            0            0          1            0
#> 202            0            0          1            0
#> 203            0            0          0            1
#> 204            0            1          0            0
#> 205            0            0          1            0
#> 206            1            0          0            0
#> 207            0            0          0            1
#> 208            0            0          1            0
#> 209            0            1          0            0
#> 210            0            0          1            0
#> 211            0            0          0            1
#> 212            0            1          0            0
#> 213            1            0          0            0
#> 214            1            0          0            0
#> 215            1            0          0            0
#> 216            0            0          1            0
#> 217            0            0          0            1
#> 218            1            0          0            0
#> 219            0            0          1            0
#> 220            0            0          1            0
#> 221            1            0          0            0
#> 222            0            0          0            1
#> 223            1            0          0            0
#> 224            1            0          0            0
#> 225            0            1          0            0
#> 226            0            0          1            0
#> 227            0            0          1            0
#> 228            0            1          0            0
#> 229            1            0          0            0
#> 230            1            0          0            0
#> 231            0            0          0            1
#> 232            0            1          0            0
#> 233            0            0          0            1
#> 234            0            1          0            0
#> 235            1            0          0            0
#> 236            1            0          0            0
#> 237            1            0          0            0
#> 238            0            0          1            0
#> 239            0            0          1            0
#> 240            0            1          0            0
#> 241            0            1          0            0
#> 242            0            0          0            1
#> 243            0            0          0            1
#> 244            0            0          0            1
#> 245            1            0          0            0
#> 246            0            0          0            1
#> 247            0            1          0            0
#> 248            0            0          1            0
#> 249            0            0          1            0
#> 250            0            1          0            0
#> 251            0            1          0            0
#> 252            0            0          0            1
#> 253            0            0          0            1
#> 254            0            1          0            0
#> 255            0            0          0            1
#> 256            0            1          0            0
#> 257            0            1          0            0
#> 258            1            0          0            0
#> 259            1            0          0            0
#> 260            0            0          0            1
#> 261            0            1          0            0
#> 262            0            0          0            1
#> 263            0            1          0            0
#> 264            0            0          0            1
#> 265            0            1          0            0
#> 266            0            0          0            1
#> 267            1            0          0            0
#> 268            1            0          0            0
#> 269            0            0          0            1
#> 270            0            0          0            1
#> 271            1            0          0            0
#> 272            0            0          0            1
#> 273            0            0          1            0
#> 274            1            0          0            0
#> 275            0            0          0            1
#> 276            0            0          0            1
#> 277            1            0          0            0
#> 278            1            0          0            0
#> 279            0            0          0            1
#> 280            0            0          0            1
#> 281            0            0          1            0
#> 282            0            0          0            1
#> 283            0            1          0            0
#> 284            0            1          0            0
#> 285            1            0          0            0
#> 286            0            0          0            1
#> 287            0            0          1            0
#> 288            1            0          0            0
#> 289            0            0          0            1
#> 290            1            0          0            0
#> 291            0            0          0            1
#> 292            0            1          0            0
#> 293            0            0          0            1
#> 294            0            0          1            0
#> 295            0            0          1            0
#> 296            1            0          0            0
#> 297            0            0          0            1
#> 298            0            1          0            0
#> 299            0            0          1            0
#> 300            1            0          0            0
#> 301            0            0          1            0
#> 302            1            0          0            0
#> 303            0            1          0            0
#> 304            0            1          0            0
#> 305            0            0          1            0
#> 306            0            0          1            0
#> 307            0            1          0            0
#> 308            0            0          1            0
#> 309            1            0          0            0
#> 310            0            1          0            0
#> 311            0            1          0            0
#> 312            0            0          0            1
#> 313            0            0          1            0
#> 314            1            0          0            0
#> 315            0            1          0            0
#> 316            0            1          0            0
#> 317            1            0          0            0
#> 318            0            0          1            0
#> 319            0            0          0            1
#> 320            0            1          0            0
#> 321            1            0          0            0
#> 322            1            0          0            0
#> 323            1            0          0            0
#> 324            0            1          0            0
#> 325            0            0          1            0
#> 326            0            0          1            0
#> 327            0            1          0            0
#> 328            0            1          0            0
#> 329            0            0          1            0
#> 330            1            0          0            0
#> 331            1            0          0            0
#> 332            0            1          0            0
#> 333            1            0          0            0
#> 334            0            0          1            0
#> 335            0            0          0            1
#> 336            0            0          1            0
#> 337            1            0          0            0
#> 338            1            0          0            0
#> 339            1            0          0            0
#> 340            0            0          1            0
#> 341            0            1          0            0
#> 342            0            0          0            1
#> 343            0            0          1            0
#> 344            0            0          0            1
#> 345            1            0          0            0
#> 346            0            1          0            0
#> 347            0            1          0            0
#> 348            1            0          0            0
#> 349            0            0          1            0
#> 350            0            0          1            0
#> 351            0            1          0            0
#> 352            1            0          0            0
#> 353            0            0          1            0
#> 354            1            0          0            0
#> 355            1            0          0            0
#> 356            0            0          0            1
#> 357            1            0          0            0
#> 358            0            0          0            1
#> 359            0            1          0            0
#> 360            0            0          0            1
#> 361            0            0          0            1
#> 362            0            0          1            0
#> 363            1            0          0            0
#> 364            0            1          0            0
#> 365            0            0          0            1
#> 366            1            0          0            0
#> 367            0            1          0            0
#> 368            0            0          0            1
#> 369            0            0          0            1
#> 370            0            0          0            1
#> 371            0            0          1            0
#> 372            0            1          0            0
#> 373            0            0          1            0
#> 374            0            0          1            0
#> 375            0            1          0            0
#> 376            0            0          1            0
#> 377            0            0          0            1
#> 378            0            0          1            0
#> 379            0            1          0            0
#> 380            1            0          0            0
#> 381            0            1          0            0
#> 382            0            1          0            0
#> 383            1            0          0            0
#> 384            0            0          0            1
#> 385            0            1          0            0
#> 386            0            0          0            1
#> 387            1            0          0            0
#> 388            0            0          0            1
#> 389            0            1          0            0
#> 390            0            0          0            1
#> 391            1            0          0            0
#> 392            0            0          0            1
#> 393            0            0          0            1
#> 394            0            0          0            1
#> 395            0            0          0            1
#> 396            1            0          0            0
#> 397            0            0          0            1
#> 398            0            0          1            0
#> 399            0            0          0            1
#> 400            0            1          0            0
#> 401            1            0          0            0
#> 402            0            1          0            0
#> 403            0            0          1            0
#> 404            0            0          0            1
#> 405            0            0          1            0
#> 406            0            0          0            1
#> 407            0            1          0            0
#> 408            0            0          1            0
#> 409            1            0          0            0
#> 410            0            1          0            0
#> 411            0            1          0            0
#> 412            1            0          0            0
#> 413            0            1          0            0
#> 414            0            1          0            0
#> 415            0            1          0            0
#> 416            0            0          1            0
#> 417            1            0          0            0
#> 418            0            1          0            0
#> 419            0            0          1            0
#> 420            1            0          0            0
#> 421            0            0          0            1
#> 422            0            0          0            1
#> 423            0            0          0            1
#> 424            0            1          0            0
#> 425            0            0          1            0
#> 426            0            1          0            0
#> 427            0            0          0            1
#> 428            0            0          0            1
#> 429            0            0          0            1
#> 430            1            0          0            0
#> 431            0            1          0            0
#> 432            0            0          1            0
#> 433            0            1          0            0
#> 434            0            1          0            0
#> 435            1            0          0            0
#> 436            1            0          0            0
#> 437            0            1          0            0
#> 438            0            0          0            1
#> 439            1            0          0            0
#> 440            1            0          0            0
#> 441            0            1          0            0
#> 442            1            0          0            0
#> 443            0            1          0            0
#> 444            1            0          0            0
#> 445            0            1          0            0
#> 446            0            0          0            1
#> 447            0            0          1            0
#> 448            0            1          0            0
#> 449            0            0          0            1
#> 450            1            0          0            0
#> 451            0            0          0            1
#> 452            0            1          0            0
#> 453            0            1          0            0
#> 454            0            1          0            0
#> 455            1            0          0            0
#> 456            0            0          1            0
#> 457            0            0          1            0
#> 458            1            0          0            0
#> 459            0            0          1            0
#> 460            0            1          0            0
#> 461            1            0          0            0
#> 462            0            0          0            1
#> 463            0            0          1            0
#> 464            1            0          0            0
#> 465            0            0          1            0
#> 466            1            0          0            0
#> 467            0            0          1            0
#> 468            0            1          0            0
#> 469            0            0          1            0
#> 470            0            1          0            0
#> 471            1            0          0            0
#> 472            0            0          1            0
#> 473            0            0          1            0
#> 474            0            0          1            0
#> 475            0            0          1            0
#> 476            0            0          1            0
#> 477            0            0          0            1
#> 478            1            0          0            0
#> 479            0            0          0            1
#> 480            0            1          0            0
#> 481            1            0          0            0
#> 482            1            0          0            0
#> 483            0            0          1            0
#> 484            1            0          0            0
#> 485            0            0          0            1
#> 486            0            0          1            0
#> 487            0            1          0            0
#> 488            0            1          0            0
#> 489            0            0          0            1
#> 490            0            1          0            0
#> 491            1            0          0            0
#> 492            1            0          0            0
#> 493            0            0          1            0
#> 494            1            0          0            0
#> 495            0            0          0            1
#> 496            0            0          1            0
#> 497            0            0          1            0
#> 498            0            0          0            1
#> 499            0            0          0            1
#> 500            1            0          0            0