Create dummy variables for all categorical/factor variables in a data.frame

Create dummy/indicator variables for all categorical variables in a data.frame. Can be used as a pre-processing step before calling other functions within the package.

Usage

dummyvars(df)

Arguments

df

a data.frame on which to make dummy variables for each categorical/factor variable, based on contrasts.

Value

a data.frame matching df but where where each categorical variable in df is replaced with indicator variables. All combinations of the indicator/dummy variable are returned. Naming convention of the new dummy variables is variable_level. For example, a factor variable in df named "colour" with levels "red", "green" and "purple" will be replaced with three columns (the new dummy variables), named colour_red, colour_green and colour_purple.

See also

Examples

dummyvars(data.frame("Colour" = factor(sample(c("red",
                                                "azure",
                                                "green",
                                                "white"),
                                             500,
                                             replace = TRUE))))
#>     Colour_azure Colour_green Colour_red Colour_white
#> 1              1            0          0            0
#> 2              0            1          0            0
#> 3              0            0          0            1
#> 4              0            1          0            0
#> 5              0            1          0            0
#> 6              0            1          0            0
#> 7              1            0          0            0
#> 8              1            0          0            0
#> 9              1            0          0            0
#> 10             0            0          1            0
#> 11             1            0          0            0
#> 12             1            0          0            0
#> 13             0            0          1            0
#> 14             0            0          0            1
#> 15             0            0          0            1
#> 16             1            0          0            0
#> 17             1            0          0            0
#> 18             0            0          1            0
#> 19             0            0          0            1
#> 20             0            0          1            0
#> 21             0            1          0            0
#> 22             1            0          0            0
#> 23             0            1          0            0
#> 24             1            0          0            0
#> 25             1            0          0            0
#> 26             0            0          0            1
#> 27             0            0          1            0
#> 28             0            1          0            0
#> 29             1            0          0            0
#> 30             0            0          1            0
#> 31             1            0          0            0
#> 32             0            0          0            1
#> 33             0            0          0            1
#> 34             0            0          1            0
#> 35             0            0          1            0
#> 36             0            0          0            1
#> 37             0            0          0            1
#> 38             1            0          0            0
#> 39             0            1          0            0
#> 40             0            0          0            1
#> 41             0            1          0            0
#> 42             1            0          0            0
#> 43             0            0          0            1
#> 44             1            0          0            0
#> 45             0            1          0            0
#> 46             0            0          0            1
#> 47             1            0          0            0
#> 48             1            0          0            0
#> 49             1            0          0            0
#> 50             0            0          1            0
#> 51             0            0          1            0
#> 52             0            0          1            0
#> 53             0            0          1            0
#> 54             0            1          0            0
#> 55             0            1          0            0
#> 56             0            0          0            1
#> 57             1            0          0            0
#> 58             1            0          0            0
#> 59             0            1          0            0
#> 60             0            0          1            0
#> 61             1            0          0            0
#> 62             0            0          1            0
#> 63             0            0          0            1
#> 64             1            0          0            0
#> 65             0            0          0            1
#> 66             0            0          1            0
#> 67             1            0          0            0
#> 68             0            0          0            1
#> 69             0            0          0            1
#> 70             1            0          0            0
#> 71             0            0          1            0
#> 72             1            0          0            0
#> 73             0            0          0            1
#> 74             1            0          0            0
#> 75             1            0          0            0
#> 76             0            0          1            0
#> 77             1            0          0            0
#> 78             0            0          1            0
#> 79             0            0          0            1
#> 80             1            0          0            0
#> 81             0            0          0            1
#> 82             0            1          0            0
#> 83             0            0          0            1
#> 84             1            0          0            0
#> 85             0            0          1            0
#> 86             0            0          1            0
#> 87             0            1          0            0
#> 88             0            0          0            1
#> 89             0            0          0            1
#> 90             0            0          1            0
#> 91             0            0          0            1
#> 92             0            0          1            0
#> 93             0            1          0            0
#> 94             0            0          1            0
#> 95             1            0          0            0
#> 96             0            1          0            0
#> 97             0            0          0            1
#> 98             0            0          0            1
#> 99             0            1          0            0
#> 100            0            1          0            0
#> 101            0            1          0            0
#> 102            0            0          1            0
#> 103            0            0          1            0
#> 104            1            0          0            0
#> 105            1            0          0            0
#> 106            0            0          1            0
#> 107            0            0          0            1
#> 108            0            1          0            0
#> 109            0            0          1            0
#> 110            0            1          0            0
#> 111            0            0          0            1
#> 112            1            0          0            0
#> 113            0            1          0            0
#> 114            0            1          0            0
#> 115            0            1          0            0
#> 116            0            0          1            0
#> 117            1            0          0            0
#> 118            0            0          1            0
#> 119            0            0          0            1
#> 120            0            0          0            1
#> 121            0            0          0            1
#> 122            1            0          0            0
#> 123            1            0          0            0
#> 124            0            0          0            1
#> 125            1            0          0            0
#> 126            1            0          0            0
#> 127            0            1          0            0
#> 128            0            0          0            1
#> 129            1            0          0            0
#> 130            0            1          0            0
#> 131            0            0          1            0
#> 132            0            0          1            0
#> 133            0            1          0            0
#> 134            1            0          0            0
#> 135            0            0          1            0
#> 136            0            0          0            1
#> 137            0            1          0            0
#> 138            1            0          0            0
#> 139            0            0          1            0
#> 140            0            1          0            0
#> 141            0            0          1            0
#> 142            1            0          0            0
#> 143            1            0          0            0
#> 144            1            0          0            0
#> 145            0            0          1            0
#> 146            0            1          0            0
#> 147            0            0          0            1
#> 148            0            0          0            1
#> 149            0            0          1            0
#> 150            0            0          1            0
#> 151            0            1          0            0
#> 152            0            1          0            0
#> 153            0            0          0            1
#> 154            1            0          0            0
#> 155            0            1          0            0
#> 156            0            0          0            1
#> 157            0            0          1            0
#> 158            0            0          0            1
#> 159            0            0          0            1
#> 160            0            0          0            1
#> 161            0            0          0            1
#> 162            0            0          1            0
#> 163            1            0          0            0
#> 164            0            1          0            0
#> 165            0            0          0            1
#> 166            0            0          1            0
#> 167            1            0          0            0
#> 168            1            0          0            0
#> 169            0            0          0            1
#> 170            0            1          0            0
#> 171            1            0          0            0
#> 172            0            0          1            0
#> 173            0            0          0            1
#> 174            0            1          0            0
#> 175            0            0          1            0
#> 176            1            0          0            0
#> 177            0            1          0            0
#> 178            1            0          0            0
#> 179            0            0          1            0
#> 180            0            1          0            0
#> 181            1            0          0            0
#> 182            0            0          1            0
#> 183            0            0          1            0
#> 184            0            0          1            0
#> 185            0            1          0            0
#> 186            1            0          0            0
#> 187            1            0          0            0
#> 188            1            0          0            0
#> 189            0            0          0            1
#> 190            1            0          0            0
#> 191            0            0          0            1
#> 192            0            1          0            0
#> 193            1            0          0            0
#> 194            0            0          0            1
#> 195            0            0          1            0
#> 196            0            0          0            1
#> 197            1            0          0            0
#> 198            1            0          0            0
#> 199            0            0          1            0
#> 200            0            0          0            1
#> 201            0            0          0            1
#> 202            0            1          0            0
#> 203            1            0          0            0
#> 204            1            0          0            0
#> 205            0            0          0            1
#> 206            0            1          0            0
#> 207            0            0          1            0
#> 208            0            0          0            1
#> 209            0            1          0            0
#> 210            0            0          0            1
#> 211            0            0          1            0
#> 212            0            0          1            0
#> 213            0            0          0            1
#> 214            0            0          1            0
#> 215            1            0          0            0
#> 216            0            0          1            0
#> 217            0            0          0            1
#> 218            0            1          0            0
#> 219            0            1          0            0
#> 220            0            1          0            0
#> 221            0            1          0            0
#> 222            0            1          0            0
#> 223            0            0          0            1
#> 224            1            0          0            0
#> 225            1            0          0            0
#> 226            0            1          0            0
#> 227            0            0          1            0
#> 228            0            1          0            0
#> 229            0            0          1            0
#> 230            0            0          0            1
#> 231            0            1          0            0
#> 232            0            0          0            1
#> 233            0            0          1            0
#> 234            1            0          0            0
#> 235            0            0          1            0
#> 236            0            1          0            0
#> 237            1            0          0            0
#> 238            0            1          0            0
#> 239            0            0          1            0
#> 240            0            1          0            0
#> 241            0            0          0            1
#> 242            0            0          0            1
#> 243            1            0          0            0
#> 244            1            0          0            0
#> 245            1            0          0            0
#> 246            0            1          0            0
#> 247            0            0          0            1
#> 248            0            0          0            1
#> 249            0            0          1            0
#> 250            0            1          0            0
#> 251            0            1          0            0
#> 252            0            0          1            0
#> 253            0            0          1            0
#> 254            0            0          0            1
#> 255            0            1          0            0
#> 256            1            0          0            0
#> 257            0            0          0            1
#> 258            0            1          0            0
#> 259            0            0          1            0
#> 260            1            0          0            0
#> 261            1            0          0            0
#> 262            0            1          0            0
#> 263            1            0          0            0
#> 264            0            0          1            0
#> 265            0            0          1            0
#> 266            0            0          0            1
#> 267            0            0          0            1
#> 268            0            0          0            1
#> 269            0            0          1            0
#> 270            1            0          0            0
#> 271            0            0          1            0
#> 272            0            0          1            0
#> 273            1            0          0            0
#> 274            0            0          1            0
#> 275            0            0          0            1
#> 276            0            1          0            0
#> 277            1            0          0            0
#> 278            0            0          1            0
#> 279            1            0          0            0
#> 280            0            1          0            0
#> 281            0            0          0            1
#> 282            1            0          0            0
#> 283            0            0          0            1
#> 284            0            1          0            0
#> 285            0            1          0            0
#> 286            0            1          0            0
#> 287            0            1          0            0
#> 288            0            0          1            0
#> 289            0            0          0            1
#> 290            0            0          1            0
#> 291            0            0          0            1
#> 292            0            0          0            1
#> 293            1            0          0            0
#> 294            0            0          1            0
#> 295            0            0          1            0
#> 296            0            1          0            0
#> 297            0            0          1            0
#> 298            0            0          1            0
#> 299            1            0          0            0
#> 300            1            0          0            0
#> 301            0            0          1            0
#> 302            0            1          0            0
#> 303            0            0          0            1
#> 304            0            0          0            1
#> 305            1            0          0            0
#> 306            0            0          1            0
#> 307            0            1          0            0
#> 308            1            0          0            0
#> 309            0            0          1            0
#> 310            1            0          0            0
#> 311            0            0          1            0
#> 312            0            0          0            1
#> 313            0            1          0            0
#> 314            0            0          1            0
#> 315            0            1          0            0
#> 316            1            0          0            0
#> 317            1            0          0            0
#> 318            1            0          0            0
#> 319            0            0          0            1
#> 320            0            0          1            0
#> 321            0            0          0            1
#> 322            0            0          1            0
#> 323            0            1          0            0
#> 324            0            0          0            1
#> 325            1            0          0            0
#> 326            1            0          0            0
#> 327            0            1          0            0
#> 328            1            0          0            0
#> 329            0            1          0            0
#> 330            0            0          1            0
#> 331            1            0          0            0
#> 332            1            0          0            0
#> 333            0            1          0            0
#> 334            0            1          0            0
#> 335            0            1          0            0
#> 336            0            1          0            0
#> 337            0            0          0            1
#> 338            0            1          0            0
#> 339            0            0          0            1
#> 340            0            0          0            1
#> 341            0            0          0            1
#> 342            1            0          0            0
#> 343            0            1          0            0
#> 344            0            0          1            0
#> 345            0            0          0            1
#> 346            0            1          0            0
#> 347            0            1          0            0
#> 348            0            0          0            1
#> 349            0            1          0            0
#> 350            0            0          1            0
#> 351            0            0          0            1
#> 352            0            0          0            1
#> 353            0            0          1            0
#> 354            0            0          1            0
#> 355            0            1          0            0
#> 356            1            0          0            0
#> 357            0            0          0            1
#> 358            1            0          0            0
#> 359            1            0          0            0
#> 360            0            0          1            0
#> 361            1            0          0            0
#> 362            0            0          0            1
#> 363            0            0          1            0
#> 364            0            1          0            0
#> 365            1            0          0            0
#> 366            0            0          0            1
#> 367            0            1          0            0
#> 368            0            0          0            1
#> 369            1            0          0            0
#> 370            0            0          1            0
#> 371            0            1          0            0
#> 372            0            0          1            0
#> 373            1            0          0            0
#> 374            0            0          1            0
#> 375            0            0          1            0
#> 376            0            1          0            0
#> 377            1            0          0            0
#> 378            0            1          0            0
#> 379            0            0          1            0
#> 380            1            0          0            0
#> 381            0            0          1            0
#> 382            0            1          0            0
#> 383            0            0          0            1
#> 384            0            0          1            0
#> 385            0            1          0            0
#> 386            0            0          0            1
#> 387            0            0          0            1
#> 388            0            1          0            0
#> 389            0            1          0            0
#> 390            0            0          1            0
#> 391            0            1          0            0
#> 392            0            0          0            1
#> 393            0            0          1            0
#> 394            0            0          1            0
#> 395            0            0          0            1
#> 396            0            0          1            0
#> 397            0            1          0            0
#> 398            0            0          0            1
#> 399            0            0          0            1
#> 400            1            0          0            0
#> 401            0            0          0            1
#> 402            0            0          1            0
#> 403            0            0          0            1
#> 404            0            0          1            0
#> 405            0            1          0            0
#> 406            0            1          0            0
#> 407            0            1          0            0
#> 408            1            0          0            0
#> 409            1            0          0            0
#> 410            0            1          0            0
#> 411            0            0          1            0
#> 412            0            1          0            0
#> 413            1            0          0            0
#> 414            0            1          0            0
#> 415            0            0          1            0
#> 416            0            1          0            0
#> 417            1            0          0            0
#> 418            0            0          1            0
#> 419            0            0          0            1
#> 420            0            0          1            0
#> 421            0            0          1            0
#> 422            0            0          1            0
#> 423            0            1          0            0
#> 424            1            0          0            0
#> 425            0            0          1            0
#> 426            0            0          0            1
#> 427            0            0          0            1
#> 428            1            0          0            0
#> 429            0            0          0            1
#> 430            0            1          0            0
#> 431            0            1          0            0
#> 432            0            0          0            1
#> 433            0            1          0            0
#> 434            1            0          0            0
#> 435            0            0          0            1
#> 436            1            0          0            0
#> 437            0            1          0            0
#> 438            0            0          0            1
#> 439            0            1          0            0
#> 440            0            0          0            1
#> 441            1            0          0            0
#> 442            1            0          0            0
#> 443            0            0          1            0
#> 444            0            0          1            0
#> 445            0            0          0            1
#> 446            0            0          0            1
#> 447            0            1          0            0
#> 448            0            0          0            1
#> 449            0            0          1            0
#> 450            0            1          0            0
#> 451            0            0          1            0
#> 452            0            0          0            1
#> 453            0            0          1            0
#> 454            0            1          0            0
#> 455            0            0          0            1
#> 456            0            0          0            1
#> 457            1            0          0            0
#> 458            0            0          1            0
#> 459            1            0          0            0
#> 460            0            1          0            0
#> 461            0            0          0            1
#> 462            0            1          0            0
#> 463            0            0          0            1
#> 464            0            1          0            0
#> 465            0            1          0            0
#> 466            1            0          0            0
#> 467            0            0          1            0
#> 468            0            1          0            0
#> 469            0            1          0            0
#> 470            1            0          0            0
#> 471            0            0          1            0
#> 472            1            0          0            0
#> 473            0            0          1            0
#> 474            0            0          0            1
#> 475            0            1          0            0
#> 476            0            0          0            1
#> 477            0            0          0            1
#> 478            0            0          0            1
#> 479            0            0          0            1
#> 480            0            0          1            0
#> 481            0            1          0            0
#> 482            0            0          0            1
#> 483            0            1          0            0
#> 484            0            0          1            0
#> 485            0            0          1            0
#> 486            0            0          1            0
#> 487            1            0          0            0
#> 488            1            0          0            0
#> 489            0            1          0            0
#> 490            0            0          0            1
#> 491            0            1          0            0
#> 492            1            0          0            0
#> 493            0            1          0            0
#> 494            0            0          0            1
#> 495            1            0          0            0
#> 496            1            0          0            0
#> 497            0            1          0            0
#> 498            0            0          0            1
#> 499            0            1          0            0
#> 500            0            0          0            1