predRupdate: Prediction Model Validation and Updating

Evaluate the predictive performance of an existing (i.e. previously developed) prediction/ prognostic model given relevant information about the existing prediction model (e.g. coefficients) and a new dataset. Provides a range of model updating methods that help tailor the existing model to the new dataset; see Su et al. (2018) doi:10.1177/0962280215626466 . Techniques to aggregate multiple existing prediction models on the new data are also provided; see Debray et al. (2014) doi:10.1002/sim.6080 and Martin et al. (2018) doi:10.1002/sim.7586 ).

See also

Author

Maintainer: Glen P. Martin glen.martin31@gmail.com (ORCID) [copyright holder]

Authors:

David Jenkins david.jenkins-5@manchester.ac.uk [copyright holder]

Matthew Sperrin matthew.sperrin@manchester.ac.uk [copyright holder]