Create dummy variables for all categorical/factor variables in a data.frame

Create dummy/indicator variables for all categorical variables in a data.frame. Can be used as a pre-processing step before calling other functions within the package.

Usage

dummy_vars(df)

Arguments

df

a data.frame on which to make dummy variables for each categorical/factor variable, based on contrasts.

Value

a data.frame matching df but where each categorical variable in df is replaced with indicator variables. All combinations of the indicator/dummy variable are returned. Naming convention of the new dummy variables is variable_level. For example, a factor variable in df named "colour" with levels "red", "green" and "purple" will be replaced with three columns (the new dummy variables), named colour_red, colour_green and colour_purple.

See also

Examples

dummy_vars(data.frame("Colour" = factor(sample(c("red",
                                                 "azure",
                                                 "green",
                                                 "white"),
                                              500,
                                              replace = TRUE))))
#>     Colour_azure Colour_green Colour_red Colour_white
#> 1              1            0          0            0
#> 2              0            0          0            1
#> 3              0            0          1            0
#> 4              0            0          0            1
#> 5              0            1          0            0
#> 6              0            1          0            0
#> 7              0            0          0            1
#> 8              0            0          0            1
#> 9              0            0          1            0
#> 10             0            0          1            0
#> 11             0            0          0            1
#> 12             0            1          0            0
#> 13             0            1          0            0
#> 14             0            0          0            1
#> 15             0            0          0            1
#> 16             0            1          0            0
#> 17             1            0          0            0
#> 18             0            1          0            0
#> 19             0            0          1            0
#> 20             0            0          0            1
#> 21             0            0          0            1
#> 22             0            1          0            0
#> 23             0            1          0            0
#> 24             0            1          0            0
#> 25             0            0          1            0
#> 26             0            0          1            0
#> 27             0            0          0            1
#> 28             1            0          0            0
#> 29             1            0          0            0
#> 30             0            0          0            1
#> 31             0            0          0            1
#> 32             1            0          0            0
#> 33             0            0          1            0
#> 34             0            1          0            0
#> 35             0            0          1            0
#> 36             1            0          0            0
#> 37             1            0          0            0
#> 38             1            0          0            0
#> 39             0            0          0            1
#> 40             0            0          0            1
#> 41             0            0          1            0
#> 42             0            0          0            1
#> 43             0            1          0            0
#> 44             0            0          0            1
#> 45             0            0          0            1
#> 46             1            0          0            0
#> 47             0            0          0            1
#> 48             0            1          0            0
#> 49             0            0          1            0
#> 50             0            1          0            0
#> 51             0            0          0            1
#> 52             0            1          0            0
#> 53             0            1          0            0
#> 54             1            0          0            0
#> 55             0            1          0            0
#> 56             0            1          0            0
#> 57             1            0          0            0
#> 58             0            0          0            1
#> 59             0            0          1            0
#> 60             1            0          0            0
#> 61             0            0          1            0
#> 62             1            0          0            0
#> 63             0            0          0            1
#> 64             0            0          1            0
#> 65             0            0          1            0
#> 66             1            0          0            0
#> 67             0            1          0            0
#> 68             0            1          0            0
#> 69             0            0          0            1
#> 70             0            1          0            0
#> 71             0            1          0            0
#> 72             1            0          0            0
#> 73             1            0          0            0
#> 74             0            0          0            1
#> 75             0            0          0            1
#> 76             0            0          0            1
#> 77             0            0          1            0
#> 78             1            0          0            0
#> 79             0            0          1            0
#> 80             0            1          0            0
#> 81             0            0          0            1
#> 82             0            0          0            1
#> 83             0            0          1            0
#> 84             0            0          1            0
#> 85             0            0          0            1
#> 86             0            0          0            1
#> 87             0            0          0            1
#> 88             0            0          1            0
#> 89             0            0          1            0
#> 90             0            1          0            0
#> 91             0            0          0            1
#> 92             0            1          0            0
#> 93             0            1          0            0
#> 94             1            0          0            0
#> 95             0            0          0            1
#> 96             0            0          1            0
#> 97             1            0          0            0
#> 98             0            0          0            1
#> 99             1            0          0            0
#> 100            0            1          0            0
#> 101            0            0          0            1
#> 102            1            0          0            0
#> 103            1            0          0            0
#> 104            0            0          0            1
#> 105            0            0          0            1
#> 106            0            0          1            0
#> 107            1            0          0            0
#> 108            0            1          0            0
#> 109            0            0          0            1
#> 110            0            0          0            1
#> 111            0            0          0            1
#> 112            1            0          0            0
#> 113            0            1          0            0
#> 114            0            0          1            0
#> 115            0            0          1            0
#> 116            0            1          0            0
#> 117            0            1          0            0
#> 118            1            0          0            0
#> 119            1            0          0            0
#> 120            0            1          0            0
#> 121            1            0          0            0
#> 122            1            0          0            0
#> 123            0            0          0            1
#> 124            1            0          0            0
#> 125            0            0          1            0
#> 126            0            1          0            0
#> 127            1            0          0            0
#> 128            1            0          0            0
#> 129            1            0          0            0
#> 130            0            0          0            1
#> 131            0            1          0            0
#> 132            0            1          0            0
#> 133            0            0          0            1
#> 134            0            0          1            0
#> 135            1            0          0            0
#> 136            0            0          1            0
#> 137            0            0          0            1
#> 138            0            0          0            1
#> 139            0            1          0            0
#> 140            0            1          0            0
#> 141            1            0          0            0
#> 142            0            0          1            0
#> 143            1            0          0            0
#> 144            0            0          1            0
#> 145            1            0          0            0
#> 146            0            0          1            0
#> 147            0            1          0            0
#> 148            1            0          0            0
#> 149            0            0          0            1
#> 150            1            0          0            0
#> 151            1            0          0            0
#> 152            0            0          0            1
#> 153            0            1          0            0
#> 154            0            1          0            0
#> 155            0            0          0            1
#> 156            0            0          1            0
#> 157            0            1          0            0
#> 158            0            1          0            0
#> 159            1            0          0            0
#> 160            0            1          0            0
#> 161            1            0          0            0
#> 162            0            1          0            0
#> 163            0            1          0            0
#> 164            0            1          0            0
#> 165            1            0          0            0
#> 166            1            0          0            0
#> 167            1            0          0            0
#> 168            0            0          1            0
#> 169            0            0          0            1
#> 170            0            1          0            0
#> 171            0            0          0            1
#> 172            0            0          1            0
#> 173            0            1          0            0
#> 174            0            0          0            1
#> 175            1            0          0            0
#> 176            1            0          0            0
#> 177            0            0          1            0
#> 178            0            0          1            0
#> 179            1            0          0            0
#> 180            1            0          0            0
#> 181            0            0          1            0
#> 182            0            1          0            0
#> 183            1            0          0            0
#> 184            0            0          1            0
#> 185            1            0          0            0
#> 186            0            1          0            0
#> 187            0            0          0            1
#> 188            0            1          0            0
#> 189            1            0          0            0
#> 190            0            0          1            0
#> 191            0            0          0            1
#> 192            1            0          0            0
#> 193            0            0          0            1
#> 194            0            0          0            1
#> 195            0            0          1            0
#> 196            0            0          1            0
#> 197            1            0          0            0
#> 198            0            0          0            1
#> 199            0            0          0            1
#> 200            0            1          0            0
#> 201            0            0          0            1
#> 202            1            0          0            0
#> 203            0            0          1            0
#> 204            0            0          0            1
#> 205            0            0          1            0
#> 206            0            0          1            0
#> 207            0            1          0            0
#> 208            0            1          0            0
#> 209            0            1          0            0
#> 210            0            1          0            0
#> 211            1            0          0            0
#> 212            0            0          0            1
#> 213            0            1          0            0
#> 214            0            0          1            0
#> 215            1            0          0            0
#> 216            1            0          0            0
#> 217            1            0          0            0
#> 218            0            1          0            0
#> 219            1            0          0            0
#> 220            0            1          0            0
#> 221            0            0          1            0
#> 222            1            0          0            0
#> 223            0            0          0            1
#> 224            0            0          1            0
#> 225            0            0          0            1
#> 226            0            1          0            0
#> 227            0            0          1            0
#> 228            0            0          1            0
#> 229            0            1          0            0
#> 230            0            1          0            0
#> 231            0            0          0            1
#> 232            0            0          1            0
#> 233            0            0          0            1
#> 234            0            0          0            1
#> 235            1            0          0            0
#> 236            0            0          1            0
#> 237            0            0          1            0
#> 238            0            0          1            0
#> 239            1            0          0            0
#> 240            1            0          0            0
#> 241            1            0          0            0
#> 242            0            0          1            0
#> 243            0            0          0            1
#> 244            1            0          0            0
#> 245            0            1          0            0
#> 246            0            1          0            0
#> 247            0            1          0            0
#> 248            0            0          1            0
#> 249            0            0          1            0
#> 250            0            0          1            0
#> 251            0            0          0            1
#> 252            0            0          0            1
#> 253            0            1          0            0
#> 254            0            1          0            0
#> 255            1            0          0            0
#> 256            0            0          1            0
#> 257            1            0          0            0
#> 258            0            0          1            0
#> 259            1            0          0            0
#> 260            0            0          1            0
#> 261            0            0          0            1
#> 262            0            1          0            0
#> 263            1            0          0            0
#> 264            0            0          0            1
#> 265            0            1          0            0
#> 266            0            0          1            0
#> 267            0            1          0            0
#> 268            1            0          0            0
#> 269            0            0          1            0
#> 270            0            1          0            0
#> 271            0            0          0            1
#> 272            0            1          0            0
#> 273            1            0          0            0
#> 274            0            1          0            0
#> 275            0            0          1            0
#> 276            0            0          1            0
#> 277            1            0          0            0
#> 278            0            0          1            0
#> 279            0            0          1            0
#> 280            1            0          0            0
#> 281            0            1          0            0
#> 282            0            1          0            0
#> 283            0            0          1            0
#> 284            0            0          1            0
#> 285            0            0          1            0
#> 286            1            0          0            0
#> 287            0            0          1            0
#> 288            1            0          0            0
#> 289            0            0          1            0
#> 290            0            1          0            0
#> 291            0            0          0            1
#> 292            0            0          1            0
#> 293            0            0          0            1
#> 294            1            0          0            0
#> 295            0            0          1            0
#> 296            0            0          1            0
#> 297            0            1          0            0
#> 298            1            0          0            0
#> 299            0            0          0            1
#> 300            0            0          1            0
#> 301            0            0          0            1
#> 302            1            0          0            0
#> 303            0            1          0            0
#> 304            0            0          0            1
#> 305            0            0          0            1
#> 306            0            0          1            0
#> 307            0            1          0            0
#> 308            1            0          0            0
#> 309            0            0          1            0
#> 310            0            0          1            0
#> 311            0            1          0            0
#> 312            0            1          0            0
#> 313            0            0          1            0
#> 314            0            0          1            0
#> 315            1            0          0            0
#> 316            0            0          1            0
#> 317            0            1          0            0
#> 318            0            0          1            0
#> 319            0            0          0            1
#> 320            0            0          0            1
#> 321            0            0          1            0
#> 322            0            1          0            0
#> 323            1            0          0            0
#> 324            0            0          0            1
#> 325            0            1          0            0
#> 326            0            0          1            0
#> 327            1            0          0            0
#> 328            0            1          0            0
#> 329            0            0          0            1
#> 330            0            1          0            0
#> 331            1            0          0            0
#> 332            0            0          0            1
#> 333            1            0          0            0
#> 334            0            0          0            1
#> 335            0            1          0            0
#> 336            0            0          1            0
#> 337            0            0          1            0
#> 338            1            0          0            0
#> 339            0            0          0            1
#> 340            0            0          1            0
#> 341            0            1          0            0
#> 342            0            0          0            1
#> 343            0            0          1            0
#> 344            0            0          0            1
#> 345            0            0          0            1
#> 346            0            0          0            1
#> 347            1            0          0            0
#> 348            1            0          0            0
#> 349            0            1          0            0
#> 350            0            0          0            1
#> 351            0            0          1            0
#> 352            0            1          0            0
#> 353            0            1          0            0
#> 354            0            0          0            1
#> 355            0            0          1            0
#> 356            0            1          0            0
#> 357            0            0          0            1
#> 358            1            0          0            0
#> 359            1            0          0            0
#> 360            0            1          0            0
#> 361            1            0          0            0
#> 362            0            0          1            0
#> 363            0            1          0            0
#> 364            0            0          1            0
#> 365            0            0          1            0
#> 366            0            1          0            0
#> 367            0            0          0            1
#> 368            1            0          0            0
#> 369            0            0          0            1
#> 370            0            0          0            1
#> 371            0            0          1            0
#> 372            0            0          1            0
#> 373            0            1          0            0
#> 374            0            0          0            1
#> 375            0            0          1            0
#> 376            0            0          0            1
#> 377            1            0          0            0
#> 378            0            0          0            1
#> 379            0            1          0            0
#> 380            0            0          0            1
#> 381            1            0          0            0
#> 382            0            0          0            1
#> 383            0            1          0            0
#> 384            0            0          0            1
#> 385            0            1          0            0
#> 386            0            0          0            1
#> 387            1            0          0            0
#> 388            0            0          1            0
#> 389            0            1          0            0
#> 390            0            0          1            0
#> 391            0            0          1            0
#> 392            0            1          0            0
#> 393            0            0          1            0
#> 394            1            0          0            0
#> 395            0            0          1            0
#> 396            0            1          0            0
#> 397            0            0          0            1
#> 398            0            1          0            0
#> 399            0            0          1            0
#> 400            0            1          0            0
#> 401            0            0          1            0
#> 402            0            1          0            0
#> 403            0            1          0            0
#> 404            0            0          0            1
#> 405            0            1          0            0
#> 406            0            0          0            1
#> 407            0            0          1            0
#> 408            0            0          1            0
#> 409            0            0          1            0
#> 410            1            0          0            0
#> 411            0            1          0            0
#> 412            0            0          1            0
#> 413            1            0          0            0
#> 414            1            0          0            0
#> 415            1            0          0            0
#> 416            0            0          1            0
#> 417            1            0          0            0
#> 418            1            0          0            0
#> 419            0            0          1            0
#> 420            0            0          1            0
#> 421            0            0          0            1
#> 422            0            0          0            1
#> 423            0            1          0            0
#> 424            1            0          0            0
#> 425            0            0          1            0
#> 426            1            0          0            0
#> 427            0            1          0            0
#> 428            0            0          1            0
#> 429            1            0          0            0
#> 430            0            1          0            0
#> 431            1            0          0            0
#> 432            1            0          0            0
#> 433            0            1          0            0
#> 434            0            0          1            0
#> 435            0            0          1            0
#> 436            0            1          0            0
#> 437            1            0          0            0
#> 438            0            1          0            0
#> 439            1            0          0            0
#> 440            0            0          1            0
#> 441            0            0          1            0
#> 442            0            0          0            1
#> 443            1            0          0            0
#> 444            0            1          0            0
#> 445            1            0          0            0
#> 446            0            0          0            1
#> 447            0            0          1            0
#> 448            0            1          0            0
#> 449            0            1          0            0
#> 450            0            0          0            1
#> 451            0            1          0            0
#> 452            1            0          0            0
#> 453            0            0          1            0
#> 454            0            0          1            0
#> 455            0            1          0            0
#> 456            0            0          0            1
#> 457            0            0          1            0
#> 458            1            0          0            0
#> 459            0            0          0            1
#> 460            0            0          0            1
#> 461            0            0          1            0
#> 462            1            0          0            0
#> 463            0            0          1            0
#> 464            1            0          0            0
#> 465            0            1          0            0
#> 466            0            0          0            1
#> 467            1            0          0            0
#> 468            0            0          1            0
#> 469            0            1          0            0
#> 470            0            1          0            0
#> 471            0            1          0            0
#> 472            1            0          0            0
#> 473            0            1          0            0
#> 474            0            0          1            0
#> 475            0            0          0            1
#> 476            0            0          0            1
#> 477            0            0          1            0
#> 478            0            0          1            0
#> 479            1            0          0            0
#> 480            0            0          1            0
#> 481            0            1          0            0
#> 482            0            0          0            1
#> 483            0            0          1            0
#> 484            1            0          0            0
#> 485            0            0          1            0
#> 486            0            1          0            0
#> 487            1            0          0            0
#> 488            0            0          1            0
#> 489            0            1          0            0
#> 490            1            0          0            0
#> 491            1            0          0            0
#> 492            0            0          1            0
#> 493            0            0          0            1
#> 494            0            0          0            1
#> 495            0            0          0            1
#> 496            0            1          0            0
#> 497            1            0          0            0
#> 498            0            1          0            0
#> 499            1            0          0            0
#> 500            0            1          0            0