Create dummy variables for all categorical/factor variables in a data.frame

Create dummy/indicator variables for all categorical variables in a data.frame. Can be used as a pre-processing step before calling other functions within the package.

Usage

dummy_vars(df)

Arguments

df

a data.frame on which to make dummy variables for each categorical/factor variable, based on contrasts.

Value

a data.frame matching df but where where each categorical variable in df is replaced with indicator variables. All combinations of the indicator/dummy variable are returned. Naming convention of the new dummy variables is variable_level. For example, a factor variable in df named "colour" with levels "red", "green" and "purple" will be replaced with three columns (the new dummy variables), named colour_red, colour_green and colour_purple.

See also

Examples

dummy_vars(data.frame("Colour" = factor(sample(c("red",
                                                 "azure",
                                                 "green",
                                                 "white"),
                                              500,
                                              replace = TRUE))))
#>     Colour_azure Colour_green Colour_red Colour_white
#> 1              0            0          0            1
#> 2              0            0          1            0
#> 3              0            1          0            0
#> 4              0            0          0            1
#> 5              1            0          0            0
#> 6              0            0          0            1
#> 7              1            0          0            0
#> 8              0            0          0            1
#> 9              0            1          0            0
#> 10             0            0          1            0
#> 11             0            0          0            1
#> 12             0            0          0            1
#> 13             0            1          0            0
#> 14             0            0          1            0
#> 15             0            1          0            0
#> 16             1            0          0            0
#> 17             0            1          0            0
#> 18             0            0          0            1
#> 19             0            0          0            1
#> 20             0            0          1            0
#> 21             0            0          0            1
#> 22             0            1          0            0
#> 23             0            1          0            0
#> 24             0            1          0            0
#> 25             1            0          0            0
#> 26             0            0          1            0
#> 27             0            0          0            1
#> 28             0            0          0            1
#> 29             0            0          1            0
#> 30             0            1          0            0
#> 31             0            0          0            1
#> 32             1            0          0            0
#> 33             0            0          0            1
#> 34             1            0          0            0
#> 35             0            1          0            0
#> 36             0            1          0            0
#> 37             1            0          0            0
#> 38             0            0          1            0
#> 39             1            0          0            0
#> 40             0            0          1            0
#> 41             0            0          0            1
#> 42             0            1          0            0
#> 43             0            0          1            0
#> 44             0            0          0            1
#> 45             1            0          0            0
#> 46             0            1          0            0
#> 47             1            0          0            0
#> 48             0            0          0            1
#> 49             0            0          0            1
#> 50             1            0          0            0
#> 51             1            0          0            0
#> 52             1            0          0            0
#> 53             0            1          0            0
#> 54             0            0          0            1
#> 55             0            0          0            1
#> 56             1            0          0            0
#> 57             1            0          0            0
#> 58             0            1          0            0
#> 59             1            0          0            0
#> 60             1            0          0            0
#> 61             0            0          0            1
#> 62             1            0          0            0
#> 63             0            0          1            0
#> 64             0            0          1            0
#> 65             0            1          0            0
#> 66             0            1          0            0
#> 67             0            0          0            1
#> 68             1            0          0            0
#> 69             1            0          0            0
#> 70             1            0          0            0
#> 71             0            0          1            0
#> 72             1            0          0            0
#> 73             0            1          0            0
#> 74             0            0          1            0
#> 75             0            0          1            0
#> 76             0            0          0            1
#> 77             0            1          0            0
#> 78             0            0          1            0
#> 79             0            1          0            0
#> 80             0            0          0            1
#> 81             0            1          0            0
#> 82             0            0          0            1
#> 83             0            1          0            0
#> 84             0            1          0            0
#> 85             0            1          0            0
#> 86             0            0          0            1
#> 87             1            0          0            0
#> 88             0            1          0            0
#> 89             0            1          0            0
#> 90             0            1          0            0
#> 91             0            0          0            1
#> 92             0            0          1            0
#> 93             0            0          1            0
#> 94             0            0          0            1
#> 95             0            0          0            1
#> 96             0            0          1            0
#> 97             1            0          0            0
#> 98             0            1          0            0
#> 99             0            1          0            0
#> 100            0            0          0            1
#> 101            0            0          1            0
#> 102            0            1          0            0
#> 103            0            0          1            0
#> 104            0            1          0            0
#> 105            0            0          0            1
#> 106            1            0          0            0
#> 107            1            0          0            0
#> 108            0            0          0            1
#> 109            0            1          0            0
#> 110            0            1          0            0
#> 111            0            1          0            0
#> 112            0            0          0            1
#> 113            0            1          0            0
#> 114            0            1          0            0
#> 115            0            0          1            0
#> 116            1            0          0            0
#> 117            0            0          1            0
#> 118            0            0          0            1
#> 119            1            0          0            0
#> 120            0            0          1            0
#> 121            1            0          0            0
#> 122            0            0          0            1
#> 123            0            1          0            0
#> 124            0            0          0            1
#> 125            1            0          0            0
#> 126            0            0          0            1
#> 127            0            1          0            0
#> 128            1            0          0            0
#> 129            0            0          0            1
#> 130            0            1          0            0
#> 131            0            0          1            0
#> 132            1            0          0            0
#> 133            1            0          0            0
#> 134            0            0          1            0
#> 135            0            1          0            0
#> 136            1            0          0            0
#> 137            1            0          0            0
#> 138            1            0          0            0
#> 139            1            0          0            0
#> 140            0            1          0            0
#> 141            0            1          0            0
#> 142            1            0          0            0
#> 143            1            0          0            0
#> 144            1            0          0            0
#> 145            0            0          1            0
#> 146            1            0          0            0
#> 147            0            0          1            0
#> 148            1            0          0            0
#> 149            0            0          1            0
#> 150            0            1          0            0
#> 151            0            1          0            0
#> 152            0            0          0            1
#> 153            0            0          0            1
#> 154            0            1          0            0
#> 155            1            0          0            0
#> 156            0            1          0            0
#> 157            0            1          0            0
#> 158            0            0          0            1
#> 159            0            0          0            1
#> 160            1            0          0            0
#> 161            0            0          0            1
#> 162            0            1          0            0
#> 163            0            0          0            1
#> 164            0            0          0            1
#> 165            0            0          0            1
#> 166            1            0          0            0
#> 167            0            1          0            0
#> 168            1            0          0            0
#> 169            0            1          0            0
#> 170            0            0          0            1
#> 171            1            0          0            0
#> 172            1            0          0            0
#> 173            0            0          0            1
#> 174            0            0          0            1
#> 175            1            0          0            0
#> 176            0            1          0            0
#> 177            0            0          1            0
#> 178            0            0          1            0
#> 179            0            0          1            0
#> 180            0            0          0            1
#> 181            1            0          0            0
#> 182            0            0          0            1
#> 183            0            0          0            1
#> 184            0            1          0            0
#> 185            0            0          0            1
#> 186            0            0          1            0
#> 187            0            0          1            0
#> 188            0            1          0            0
#> 189            1            0          0            0
#> 190            0            0          1            0
#> 191            1            0          0            0
#> 192            0            0          0            1
#> 193            1            0          0            0
#> 194            0            1          0            0
#> 195            1            0          0            0
#> 196            0            0          0            1
#> 197            1            0          0            0
#> 198            0            0          0            1
#> 199            0            0          1            0
#> 200            0            0          0            1
#> 201            1            0          0            0
#> 202            0            0          1            0
#> 203            1            0          0            0
#> 204            0            0          0            1
#> 205            0            1          0            0
#> 206            0            0          0            1
#> 207            0            0          0            1
#> 208            0            0          1            0
#> 209            0            1          0            0
#> 210            0            0          1            0
#> 211            0            0          1            0
#> 212            0            0          1            0
#> 213            0            1          0            0
#> 214            0            0          0            1
#> 215            0            0          0            1
#> 216            0            0          0            1
#> 217            0            0          0            1
#> 218            0            1          0            0
#> 219            1            0          0            0
#> 220            0            1          0            0
#> 221            0            0          0            1
#> 222            0            0          1            0
#> 223            0            0          1            0
#> 224            0            0          1            0
#> 225            0            0          0            1
#> 226            0            1          0            0
#> 227            0            1          0            0
#> 228            1            0          0            0
#> 229            1            0          0            0
#> 230            0            1          0            0
#> 231            0            0          1            0
#> 232            0            1          0            0
#> 233            0            0          0            1
#> 234            0            1          0            0
#> 235            0            1          0            0
#> 236            0            0          0            1
#> 237            1            0          0            0
#> 238            0            1          0            0
#> 239            0            1          0            0
#> 240            0            0          0            1
#> 241            0            0          1            0
#> 242            1            0          0            0
#> 243            1            0          0            0
#> 244            0            0          1            0
#> 245            0            0          1            0
#> 246            0            1          0            0
#> 247            0            1          0            0
#> 248            0            1          0            0
#> 249            0            1          0            0
#> 250            0            1          0            0
#> 251            0            0          1            0
#> 252            0            1          0            0
#> 253            1            0          0            0
#> 254            0            0          1            0
#> 255            0            0          0            1
#> 256            0            0          0            1
#> 257            0            1          0            0
#> 258            0            0          0            1
#> 259            0            0          1            0
#> 260            0            1          0            0
#> 261            0            1          0            0
#> 262            0            1          0            0
#> 263            0            1          0            0
#> 264            0            1          0            0
#> 265            1            0          0            0
#> 266            0            0          1            0
#> 267            0            1          0            0
#> 268            0            0          1            0
#> 269            0            0          0            1
#> 270            1            0          0            0
#> 271            1            0          0            0
#> 272            0            0          0            1
#> 273            1            0          0            0
#> 274            0            0          1            0
#> 275            0            1          0            0
#> 276            0            1          0            0
#> 277            0            1          0            0
#> 278            0            1          0            0
#> 279            0            0          0            1
#> 280            0            1          0            0
#> 281            0            0          0            1
#> 282            0            0          1            0
#> 283            0            0          0            1
#> 284            0            1          0            0
#> 285            1            0          0            0
#> 286            0            0          0            1
#> 287            0            0          1            0
#> 288            0            0          1            0
#> 289            0            0          1            0
#> 290            0            0          0            1
#> 291            0            0          1            0
#> 292            0            0          1            0
#> 293            0            1          0            0
#> 294            0            0          0            1
#> 295            1            0          0            0
#> 296            1            0          0            0
#> 297            0            0          0            1
#> 298            0            0          0            1
#> 299            0            1          0            0
#> 300            0            0          0            1
#> 301            0            0          0            1
#> 302            0            0          1            0
#> 303            0            0          1            0
#> 304            0            1          0            0
#> 305            0            0          0            1
#> 306            0            1          0            0
#> 307            0            1          0            0
#> 308            0            0          1            0
#> 309            1            0          0            0
#> 310            1            0          0            0
#> 311            0            0          0            1
#> 312            1            0          0            0
#> 313            1            0          0            0
#> 314            1            0          0            0
#> 315            0            0          0            1
#> 316            1            0          0            0
#> 317            1            0          0            0
#> 318            1            0          0            0
#> 319            0            0          0            1
#> 320            0            0          0            1
#> 321            1            0          0            0
#> 322            1            0          0            0
#> 323            0            0          1            0
#> 324            0            0          1            0
#> 325            0            0          1            0
#> 326            1            0          0            0
#> 327            1            0          0            0
#> 328            0            0          1            0
#> 329            0            0          0            1
#> 330            0            1          0            0
#> 331            1            0          0            0
#> 332            0            0          0            1
#> 333            0            1          0            0
#> 334            0            1          0            0
#> 335            0            0          1            0
#> 336            0            1          0            0
#> 337            0            0          0            1
#> 338            0            0          0            1
#> 339            0            0          1            0
#> 340            0            1          0            0
#> 341            0            0          0            1
#> 342            1            0          0            0
#> 343            0            0          0            1
#> 344            0            0          1            0
#> 345            1            0          0            0
#> 346            1            0          0            0
#> 347            0            0          1            0
#> 348            0            0          0            1
#> 349            0            1          0            0
#> 350            1            0          0            0
#> 351            1            0          0            0
#> 352            1            0          0            0
#> 353            0            0          0            1
#> 354            1            0          0            0
#> 355            0            0          1            0
#> 356            0            0          0            1
#> 357            0            0          0            1
#> 358            0            0          1            0
#> 359            0            1          0            0
#> 360            0            1          0            0
#> 361            0            1          0            0
#> 362            0            0          1            0
#> 363            0            0          0            1
#> 364            1            0          0            0
#> 365            0            1          0            0
#> 366            0            0          1            0
#> 367            0            0          1            0
#> 368            0            0          0            1
#> 369            0            0          1            0
#> 370            1            0          0            0
#> 371            0            0          0            1
#> 372            1            0          0            0
#> 373            0            1          0            0
#> 374            1            0          0            0
#> 375            1            0          0            0
#> 376            1            0          0            0
#> 377            0            0          1            0
#> 378            0            1          0            0
#> 379            1            0          0            0
#> 380            0            0          0            1
#> 381            1            0          0            0
#> 382            1            0          0            0
#> 383            0            0          0            1
#> 384            1            0          0            0
#> 385            0            0          0            1
#> 386            0            0          0            1
#> 387            1            0          0            0
#> 388            0            1          0            0
#> 389            0            0          0            1
#> 390            0            0          0            1
#> 391            0            1          0            0
#> 392            1            0          0            0
#> 393            0            1          0            0
#> 394            0            0          0            1
#> 395            0            1          0            0
#> 396            0            0          0            1
#> 397            0            0          1            0
#> 398            0            0          0            1
#> 399            1            0          0            0
#> 400            1            0          0            0
#> 401            0            0          0            1
#> 402            1            0          0            0
#> 403            1            0          0            0
#> 404            1            0          0            0
#> 405            1            0          0            0
#> 406            0            0          0            1
#> 407            0            0          0            1
#> 408            0            0          1            0
#> 409            1            0          0            0
#> 410            0            0          0            1
#> 411            0            0          1            0
#> 412            1            0          0            0
#> 413            0            1          0            0
#> 414            1            0          0            0
#> 415            0            0          0            1
#> 416            1            0          0            0
#> 417            0            0          0            1
#> 418            1            0          0            0
#> 419            0            1          0            0
#> 420            0            0          0            1
#> 421            0            0          0            1
#> 422            1            0          0            0
#> 423            0            0          0            1
#> 424            1            0          0            0
#> 425            1            0          0            0
#> 426            0            0          1            0
#> 427            1            0          0            0
#> 428            0            0          0            1
#> 429            0            0          0            1
#> 430            1            0          0            0
#> 431            1            0          0            0
#> 432            1            0          0            0
#> 433            0            0          0            1
#> 434            0            1          0            0
#> 435            1            0          0            0
#> 436            1            0          0            0
#> 437            0            1          0            0
#> 438            0            0          0            1
#> 439            0            1          0            0
#> 440            1            0          0            0
#> 441            1            0          0            0
#> 442            0            1          0            0
#> 443            0            0          0            1
#> 444            1            0          0            0
#> 445            1            0          0            0
#> 446            0            0          1            0
#> 447            0            0          0            1
#> 448            0            0          1            0
#> 449            0            1          0            0
#> 450            0            1          0            0
#> 451            0            0          1            0
#> 452            0            0          0            1
#> 453            0            1          0            0
#> 454            0            1          0            0
#> 455            1            0          0            0
#> 456            1            0          0            0
#> 457            0            0          1            0
#> 458            0            0          0            1
#> 459            0            0          0            1
#> 460            1            0          0            0
#> 461            0            0          1            0
#> 462            0            0          0            1
#> 463            0            0          1            0
#> 464            0            1          0            0
#> 465            0            0          0            1
#> 466            1            0          0            0
#> 467            0            1          0            0
#> 468            0            1          0            0
#> 469            0            0          0            1
#> 470            0            0          1            0
#> 471            0            0          1            0
#> 472            0            0          1            0
#> 473            0            0          1            0
#> 474            0            0          0            1
#> 475            0            0          0            1
#> 476            0            0          1            0
#> 477            1            0          0            0
#> 478            0            0          1            0
#> 479            0            1          0            0
#> 480            0            0          1            0
#> 481            0            0          0            1
#> 482            1            0          0            0
#> 483            0            0          1            0
#> 484            0            1          0            0
#> 485            0            0          0            1
#> 486            0            0          0            1
#> 487            0            0          0            1
#> 488            0            0          0            1
#> 489            0            1          0            0
#> 490            0            0          0            1
#> 491            1            0          0            0
#> 492            0            1          0            0
#> 493            1            0          0            0
#> 494            0            0          1            0
#> 495            1            0          0            0
#> 496            1            0          0            0
#> 497            1            0          0            0
#> 498            0            0          1            0
#> 499            0            0          0            1
#> 500            0            0          1            0